pyproject.toml
[tool.poetry]
name = "discordpy-bot-template-with-commands-extension"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.4"
flask = "^2.0.1"
youtube-dl = "^2021.5.16"
PyNaCl = "^1.4.0"
ffmpeg = "^1.4"
humanize = "^3.5.0"
discord-ext-slash = "^0.7.0"
discord-py-slash-command = "^1.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
